Add compatibility with Microsoft Visual C++ 2005 Express Edition Beta 2 (a complete VC++ IDE available for free). It should not break MSVC++ 2003 compatibility. (src/win32/README.vc8 will be completed later)
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@2359 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/lib/smartall.h
@@ -100,12 +100,12 @@ public:
 
 void *operator new(size_t s, const char *fname, int line)
 {
-  void *p = sm_malloc(fname, line, s > sizeof(int) ? s : sizeof(int));
+  void *p = sm_malloc(fname, line, s > sizeof(int) ? (unsigned int)s : sizeof(int));
   return p;
 }
 void *operator new[](size_t s, const char *fname, int line)
 {
-   void *p = sm_malloc(fname, line, s > sizeof(int) ? s : sizeof(int));
+   void *p = sm_malloc(fname, line, s > sizeof(int) ? (unsigned int)s : sizeof(int));
    return p;
 }
 void  operator delete(void *ptr)
